feat: orderFruit 함수 생성, Enum 파일 추가
- orderFruit: 과일 재고 추가하는 함수
