+++ JuiceMaker/JuiceMaker/Model/JuiceMaker.swift
@@ -6,26 +6,16 @@
 
 import Foundation
 
-enum Juice {
-    case strawberry, banana, kiwi, pineapple, strawNana, mango, mangKi
-}
-
-enum JuiceError: Error {
-    case insufficientFruitStock
-    case notFoundFruitInformation
-}
-
-// 쥬스 메이커 타입
 struct JuiceMaker {
     private let fruitStore = FruitStore()
 
     private func useFruits(_ fruit: Fruit, _ amount: Int) throws {
         guard let stock = fruitStore.checkStock(of: fruit) else {
-            throw JuiceError.notFoundFruitInformation
+            throw FruitJuiceError.notFoundFruitInformation
         }
         
         if stock < amount {
-            throw JuiceError.insufficientFruitStock
+            throw FruitJuiceError.insufficientFruitStock
         }
         
         try fruitStore.consumeStock(of: fruit, by: amount)
